refactor: consolidate demo utilities and radio demo styles
demo-support.js:
- Extract shared renderPage, addCopyButtons, and openAccordion
  into packages/utils/src/demo-support.js
- Expose via @aurodesignsystem/utils/demo-support subpath export
  to keep markdown/rehype deps out of the main bundle entry point
- Replace all six component demo-support.js files with a single
  re-export from @aurodesignsystem/utils/demo-support

radio demo styles:
- Replace styles.css and inline demoWrapper CDN link with a
  compiled styles.min.css built from styles.scss
- styles.scss uses @import '@aurodesignsystem/config/demo-styles'
  matching the shared styles pattern
- Update all radio demo HTML files to reference styles.min.css
- Update package.json scripts: add docs command, update serve and
  build:docs to use the new auro docs runner
