+++ components/radio/demo/styles.scss
@@ -1,0 +1,1 @@
+@import '@aurodesignsystem/config/demo-styles';
