+++ components/radio/package.json
@@ -48,12 +48,13 @@
     "build:sass": "sass --no-source-map --load-path=../../node_modules \"./src/styles/:./src/styles/\"",
     "build:watch": "nodemon -e scss,js --watch src --ignore \"*.css\" --ignore \"*-css.js\" --exec \"npm run build\"",
     "bundler": "rollup -c node:@aurodesignsystem/config/rollup",
-    "build:docs": "wca analyze 'scripts/wca/*.js' --outFiles docs/api.md; node ../../packages/build-tools/src/docProcessor.mjs --component radio",
+    "build:docs": "wca analyze 'scripts/wca/*.js' --outFiles docs/api.md; npm run docs",
     "clean": "rm -rf dist build",
     "dev": "concurrently \"npm run build:watch\" \"npm run serve\"",
     "dev:closed": "concurrently \"npm run build:watch\" \"npm run serve:closed\"",
-    "serve": "web-dev-server --open",
-    "serve:closed": "web-dev-server",
+    "serve": "npm run docs -- -s -w --open",
+    "serve:closed": "npm run docs -- -s -w",
+    "docs": "auro docs -r '../../docs/templates/componentReadmeTemplate.md'",
     "postCss:component": "node ../../node_modules/@aurodesignsystem/auro-library/scripts/build/postCss.mjs",
     "sass:render": "sass-render --load-path=../../node_modules 'src/**/*.css' -t ../../node_modules/@aurodesignsystem/auro-library/scripts/build/staticStyles-template.js",
     "test": "wtr --config \"../../packages/config/src/web-test-runner.config.mjs\" --coverage --port 7701",
